+++ docs/whats-new.rst
@@ -8,8 +8,22 @@ v0.1.9 (X XXX. 202X)
 
 **Features and front-end API**
 
-- New methods to preprocessed data for OWC software. These methods can preprocessed Argo data in python and possibly create
-float_source/<WMO>.mat files to be used as inputs for OWC implementations in `Matlab <https://github.com/ArgoDMQC/matlab_owc>`_ and `Python <https://github.com/euroargodev/argodmqc_owc>`_.
+- New methods to preprocess data for OWC software. These methods can preprocessed Argo data in python and possibly create float_source/<WMO>.mat files to be used as inputs for OWC implementations in `Matlab <https://github.com/ArgoDMQC/matlab_owc>`_ and `Python <https://github.com/euroargodev/argodmqc_owc>`_. (:pr:`142`) by `G. Maze <http://www.github.com/gmaze>`_.
+  This new method comes with others improvements:
+
+  - New ``filter_scalib_pres`` method to filter variables according to OWC salinity calibration software requirements,
+  - New ``subsample_pressure`` method to subsample a dataset down to one value by pressure bins,
+  - Improved ``filter_qc`` method to select which fields to consider,
+  - Add conductivity (``CNDC``) to the possible output of the ``TEOS10`` method.
+
+.. code-block:: python
+
+    from argopy import DataFetcher as ArgoDataFetcher
+    ds = ArgoDataFetcher(mode='expert').float(6902766).load().data
+    ds.argo.create_float_source("float_source")
+    ds.argo.create_float_source("float_source", force='raw')
+    ds_source = ds.argo.create_float_source()
+
 
 v0.1.8 (2 Nov. 2021)
 ---------------------